backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
dnspython==2.6.1
executing==2.1.0
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.1.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
praw==7.7.1
prawcore==2.4.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
